Big fix

+++ public/js/root.js
@@ -1,9 +1,10 @@
 let del = document.querySelector('.delete-article');
-
-del.addEventListener('click', () => {
-  fetch(`http://localhost:3000/articles/del/${del.id}`, {
-    method: 'DELETE'
-  })
-    .then(res => res.text())
-    .then(data => console.log(data));
-});
+if (del) {
+  del.addEventListener('click', () => {
+    fetch(`http://localhost:3000/articles/del/${del.id}`, {
+      method: 'DELETE'
+    })
+      .then(res => res.text())
+      .then(data => console.log(data));
+  });
+}
